restructure modules for turboquant
