+++ vortex-tensor/src/encodings/turboquant/array/centroids.rs
