+++ vortex-tensor/src/encodings/turboquant/compute/take.rs
@@ -8,8 +8,8 @@ use vortex_array::IntoArray;
 use vortex_array::arrays::dict::TakeExecute;
 use vortex_error::VortexResult;
 
-use crate::encodings::turboquant::array::TurboQuant;
-use crate::encodings::turboquant::array::TurboQuantData;
+use crate::encodings::turboquant::TurboQuant;
+use crate::encodings::turboquant::TurboQuantData;
 
 impl TakeExecute for TurboQuant {
     fn take(
